requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pip==22.0.4
pycoingecko==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==58.1.0
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
